Separated parsing JavaScript code (parsing directory) from generating SQL code (generating directory).
Files in the parsing directory have the "parsing" suffix. Files in the the generation direction have the "generation" suffix.

Introduced the SelectStatement interface to collect all information required to generate the code for the SELECT statement. An object satisfying SelectStatement is initialized in the constructor of the Table class and then passed on to classes representing table-level instructions, where it is copied and extended.

SQL code is now fully generated in a hierarchy of separate functions. It is no longer generated in the classes representing table-level instructions.

Introduced PropertyOperation interfaces to represent different types of operations on the property level.

The javascript_parsing.js file now only contains parsers related to plain JavaScript code. "createDictionaryParser" now expects a parser for key/value pairs. "createObjectPropertyParser" is narrowed to objects with provided names. Property names are no longer rewritten to table names at the parsing level. The "aString" parser now supports both notations: strings surrounded by single quotes and double quotes.

Rewrote predicate parsing based on the concepts comparisons, clauses inside parentheses and concatenations.

Because queries for only table are supported in this version, the table alias "t1" is hard-coded. This will be replaced in the next version when support for joins will be implemented.

Replaced the overridden "toString" method, previously used to generate SQL code, with a "toSql" method.