Implemented basic support for a count subquery in a map with one or more filters.
Introduced the notion of a selection as a list of column-level operations. There is a generator for each type of selection. Some selection-level generators use column-level generators.

Implemented a predicate expression parser entirely based on argsecond. At this point, the parser is only used for subqueries.

A Filter interface which includes a mapping of the parameter names available in the predicate function to the tables is now wrapped around predicate expressions. This is necessary for predicates in subqueries which may refer to parameters both in the inner and the outer function.

Similarly, group keys are now represented by the Key interface rather than an array of objects satisfying PartOfKey. The Key interface includes a mapping of the parameter names used in the getKey function to the tables.

Added a roadmap.

Replaced the Alias interface with pairs where the first item represents the alias. Renamed Get to GetFromParameter.

Simplified the joins property of the SelectStatement interface from an array to a single object, because the current version only supports joins of two tables. Extracted the creation of an empty select statement to a function.

Replaced the constructor dictionary in parseSelectMultipleTables with an array to preserver order.

Renamed createObjectPropertyParser to createNamedObjectPropertyParser and added a more general new createObjectPropertyParser function that accepts an object parser as an argument.

Switching from Array<T> to the T[] notation.