Replaced COUNT(column) with COUNT(*) in grouped tables.
Introduced two types, ColumnAggregationOperation and AggregationOperations, for operations that are available for a particular column or on the level of tables/joins, respectively; refactored aggregation parsing by factoring out the creation of the key/value parser; renamed createAccessKeyParser to createGetPartOfKeyParser (in line with the GetPartOfKey interface name).

Moved the parser for a single parameterless function invocation used in createParameterlessFunctionInvocationChoice to a separate function.

Improved test titles in aggregation_parsing.test.ts.

Added the missing v0.6 entry in the changelog.