Groups can now be ordered by aggregation operations. Sorting is allowed after grouping, but disallowed before grouping. Distinguished between OrderExpression and GroupOrderExpression and between SelectStatement and GroupSelectStatement. Order expressions use GetColumn, whereas group order expressions use the same three supported column operations that are available for aggregation.
Moved the parser that a provides a choice between these operations to a separate function and (re-)used the code for group ordering and aggregation. Split up the three column aggregation parsers into separate files.

Fixed the conditions used in checkIfGetColumnInOrderIsAbsentFromGetSelection to determine the absence of a column in the ORDER BY clause from the SELECT clause.

Defined the interfaces Count, Max, Min, Avg and Sum to represent the results of applying aggregation functions. Defined StringAggregationRecord to, more precisely, constrain the result type of the function passed to the aggregate method. Renamed "aggregationOperation" in AggregateColumn to "aggregationFunction".

Added a helper function to assert that a parser does not match a given function.

Updated changelog and roadmap.