Implemented support for Boolean values and negations.
Extended predicates to cover evaluations of boolean values.

Rewrote predicate parsing tests and split them up into several files.

Added a constant predicate parser that is used when neither tables nor user-provided data are used in a predicate. Rewrote the existing two factory functions for parameterless/parameterized predicate parsers to build upon the parser for constants.

The sides of a comparison can now be in parentheses. Added tests for Side parsing.

Added the fulltime property to the Emplyoee case to test support for booleans.

Updated changelog and roadmap.