Replaced the use of constructors with column records. Removed the Constructor<T> type and the code related to parsing constructors.
Replaced employeeColumns with references to employees.columns. Added test data for companies.

Updated readme: fixed a typo in the project description, added the missing "Features" header, replaced the class in step 1 with an interface, updated the defineTable example, fixed the example of generated SQL code.

Updated readme, changelog and roadmap.