Reused the SQLite integration tests for Postgres.
Numeric column types are now specified to be "integer" or "float" (rather than "numeric"). Renamed ColumnRecord to ColumnTypeRecord.

The SQL dialect now has properties that indicate how the client returns booleans and floats. The adaptation of client results has been moved from the dialects to DatabaseContext, using these two dialect properties. For Postgres, numeric results which are returned as strings by the client are converted to floats.

Minimization and maximization operations now work on boolean columns in Postgres. To support this, basic support for conversions from booleans to integers has been added. Furthermore, the determination of result column types has been added extended from a single type to a pairing between the type as returned by the client and the type expected by the API. A maximized/minimized boolean column is converted to integer during selection and then converted back to booleans.

Added a single group aggregation operation selection. This is used when a SELECT DISTINCT statement which selects only a single column is adapted, because that single column is absent from the ORDER BY clause.

Refactored the adaptation of the SELECT statement in PostgresDialect using a reduction operation applied on rules.

Removed the types Max, Min, Avg, Sum, Count, GroupAggregationRecord and TableAggregationRecord which were used incorrectly. Set the return types of the methods "avg", "max", "min" and "sum" in StringAggregation, NumberAggregation and BooleanAggregation to primitive types. Improved AggregatableTable using type mapping.

Removed "properties" from SingleTableSelection and simplified "namesPairedWithProperties" in MultiTableSelection to a string array property named "names". Information about column in individual tables is available on the SELECT statement level and was, thus, redundant in the interfaces representing selections. Consequently, no parsing is needed for single and multi-table selections: parseSingleTableSelection and parseMultipleTableSelection have been removed.